Build fix

+++ cobalt/speech/endpointer/energy_endpointer.cc
@@ -32,7 +32,7 @@ int64_t Secs2Usecs(float seconds) {
 
 float GetDecibel(float value) {
   if (value > 1.0e-100)
-    return 20 * log10(value);
+    return static_cast<float>(20 * log10(value));
   return -2000.0;
 }
